+++ Changelogs/Changelog.md
@@ -25,3 +25,6 @@
 ### Fixed (72.3)
 - Superior (Shadowlands) and Epic (Shadowlands) are now marked as no longer obtainable
 - Watch List and Excluded categories no longer show Unknown as their names
+
+### Fixed (72.4)
+- Remix: Pandaria achievements now show the remix icon on the achievement with the duration of the event